refactor: move router types into the per-entry types modules
`createTrueSheetRouter.ts` declared `TrueSheetRouterOptions` and
`StackRouterFactory` against React Navigation directly, so the Expo Router entry
had to name React Navigation types for its `StackRouter` cast — degrading them
to `any` there.

Move both (plus `TrueSheetRouter`/`TrueSheetRouterState`) into each entry's
`types.ts`, so every base-dependent type now has exactly one home per entry and
`createTrueSheetRouter.ts` imports no base types at all. Pre-instantiate the
shared factory parameter so either entry's `StackRouterFactory` fits it without
a cast, and pin the router value at the Expo boundary.

Public names on `/navigation` are unchanged.
